27 change cart amount (#27)
* Fix issue where 'Cart' was displaying below the cart icon instead of to the right

* Implement amount change in cart

A few issues:
The view is a little janky, it feels like we don't have a lot of control over how things are laid out, so that may need to be changed at some point.
This is all using our current cart implementation. Once we change that into an activerecord object, the form will need to change to reflect that.